refactor: all commands are now in separate modules + linting and re-org
